backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
boto3==1.35.80
